folders organized and landing page started
