+++ frontend/src/components/LandingPage/LandingPage.js
@@ -1,0 +1,9 @@
+
+
+const LandingPage = () => {
+    return (
+        <h1>WELCOME</h1>
+    )
+}
+
+export default LandingPage;
