+++ frontend/src/components/auth/ProtectedRoute.js
@@ -6,7 +6,7 @@ const ProtectedRoute = props => {
   const user = useSelector(state => state.session.user)
   return (
     <Route {...props}>
-      {(user)? props.children  : <Redirect to='/login' />}
+      {(user) ? props.children : <Redirect to='/home' />}
     </Route>
   )
 };
